requirements.txt
fastapi==0.115.9
langchain==0.3.25
langchain-openai==0.3.18
python-dotenv==1.1.0
rich==14.0.0
uvicorn==0.34.2